Overhauled legendflex for better compatibility with HG2 graphics
In this update, I've made some major changes to the code so that it is
now compatibile with both old and new (post-R2014b) handle graphics.  Now the same
function can be run in any version of Matlab, rather than needing two functions
(which had been my stopgap solution since the release of R2014b).  Primary changes
involved replacing the old resize function with individual listeners to keep
the legend in sync with changing axes positions and labeled-object properties, and
updating the preallocation of handles as either numerics or graphics objects as
appropriate to version.

The contour-labeling bug is still present as of R2015a-beta... will need to keep
an eye on this as new releases come out.