Corrected bug in listeners when legend deleted; hggroup issue discovered.
I corrected a bug where the callback that resyncs the legend to the labeled
objects would cause an error if the user manually deleted the legend.  It now
verifies that the legend exists before running the listener callbacks.

Also, a user alerted me to a bug that occurs if one attempts to label an hggroup
object in R2014b+.  For some reason, the legend line/patch/etc objects have an
empty Tag property, rather than being labeled with the appropriate legend text.
I use those tags to figure out which symbols go where, so for now I do not have
a workaround for that, except to advise users not to pass hggroup objects to
legendflex.