Changed default ncol/nrow value from NaN to 0
In previous versions, if the default value for nrow/ncol input was passed
in manually, the NaN value would fail the integer validation check.  I
hadn't originally intended a user to manually pass that, but it's perfectly
reasonable to assume they might (for code consistency reasons), so I've
changed the default value to 0.  The value still means "as many as necessary".