Fixed bug in R2016a parsing of text properties
Compatibility with all the Matlab releases, which have been slowly
rolling out new legend objects and syntax between R2014b and R2016a
(with various bugs mixed in), is getting tricky.

Thus far I'm continuing to rely on the legacy version of legend
(make_legend with version='on'),
which creates an axis object, rather than a legend object.  Despite
my best attempts, and dives into the undocumented properties of the
new legend objects, I haven't found a good way to access and alter
the legend object's child text and plot marker objects.

In R2016a, the legacy version of the legend command introduces a bug
where text properties (such as font size and interpreter)
are not properly applied to the legend text.  This update includes
a workaround for this bug.

I also updated the input parsing to use the newer addParameter
inputParser method, instead of addParamValue.