Changed legend box to be visible but with no color
Implemented change that keeps the legend axis visible, but sets its
color (background, xcolor, and ycolor) to be "None".  This allows the
axis box to still be manually selected and dragged by the user after
creation.

Note that this change is only possible in newer versions of Matlab,
since HG1 graphics do not allow for axis line color to be set to
None.